+++ src/main/java/SmartFoodStreet_Backend/dto/authentication/response/RegisterResponse.java
@@ -9,7 +9,7 @@ import lombok.experimental.FieldDefaults;
 @Builder
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class RegisterResponse {
-    int accountId;
+    long accountId;
 
     String userName;
 }
